Split files utils and CustomNamedTemporaryFile to own files
This had to be done, otherwise we got circular imports with the uploads module.
I also replaced urlretrieve with requests.get because it is more reliable.
